+++ src/server/devices/baseDeviceInterfaces/iBluetoothDetector.ts
@@ -2,7 +2,7 @@ import { iRoomDevice } from './iRoomDevice';
 import { ProximityCallback, TrilaterationBasePoint } from '../espresense';
 
 export interface iBluetoothDetector extends iRoomDevice {
-  position?: TrilaterationBasePoint;
+  position: TrilaterationBasePoint;
 
   /**
    * Add callback to react on a specified device entering/leaving a given zone