+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Correct timing for Trilateration initialize call.
+
 ## 2.15.0 (2023-10-01)
 
 * (theimo1221) Add Functions for trilateration and distance calculation